Markdown file from Kokkos wiki
May need to bring internal documentation written by David Poliakoff on Kokkos Tools out front and in here. Some of it, especially the tuning interface documentation and documentation for developing a kokkos connector, can be relevant to Kokkos Application programmers (as evidenced through the Kokkos Slack in the last few months).